requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Security
Werkzeug==3.0.1
WTForms==3.1.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
